+++ packages/SUI.BASE_COMPONENTS/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.1](https://github.com/mbuyakov/SUI.CORE/compare/v2.3.0...v2.3.1) (2019-08-08)
+
+
+### Bug Fixes
+
+* **report:** fix import & dep ([a35cf6c](https://github.com/mbuyakov/SUI.CORE/commit/a35cf6c))
+
+
+
+
+
+# [2.3.0](https://github.com/mbuyakov/SUI.CORE/compare/v2.2.0...v2.3.0) (2019-08-08)
+
+
+### Features
+
+* **report:** add report components ([845a13f](https://github.com/mbuyakov/SUI.CORE/commit/845a13f))
+
+
+
+
+
 # [2.2.0](https://github.com/mbuyakov/SUI.CORE/compare/v2.1.0...v2.2.0) (2019-08-08)